pyproject.toml
[project]
name = "etl-kafka-spark-mongodb"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "kafka-python>=2.2.11",
    "pyspark>=4.0.0",
    "requests>=2.32.4",
]
